backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
decouple==0.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
PyJWT==1.7.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
